refactor(dom): move createClassNames to utils
