+++ packages/react-instantsearch-dom/src/core/utils.js
@@ -1,4 +1,18 @@
-export function isSpecialClick(event) {
+import cx from 'classnames';
+
+export const createClassNames = (block, prefix = 'ais') => (...elements) => {
+  const suitElements = elements
+    .filter(element => element || element === '')
+    .map(element => {
+      const baseClassName = `${prefix}-${block}`;
+
+      return element ? `${baseClassName}-${element}` : baseClassName;
+    });
+
+  return cx(suitElements);
+};
+
+export const isSpecialClick = event => {
   const isMiddleClick = event.button === 1;
   return Boolean(
     isMiddleClick ||
@@ -7,8 +21,7 @@ export function isSpecialClick(event) {
       event.metaKey ||
       event.shiftKey
   );
-}
+};
 
-export function capitalize(key) {
-  return key.length === 0 ? '' : `${key[0].toUpperCase()}${key.slice(1)}`;
-}
+export const capitalize = key =>
+  key.length === 0 ? '' : `${key[0].toUpperCase()}${key.slice(1)}`;
